+++ Source/Model/ModelNode.cpp
@@ -46,8 +46,8 @@ namespace EastEngine
 			});
 			m_vecMaterial.clear();
 
-			uint32_t nSize = m_vecChildModelNode.size();
-			for (uint32_t i = 0; i < nSize; ++i)
+			size_t nSize = m_vecChildModelNode.size();
+			for (size_t i = 0; i < nSize; ++i)
 			{
 				if (m_vecChildModelNode[i] != nullptr)
 				{
